requirements.txt
click==8.0.1
dataclasses==0.8
Flask==2.0.1
importlib-metadata==4.7.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
pkg-resources==0.0.0
protobuf==3.17.3
six==1.16.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
